Created an interface ILazBackend
A new interface named ILazBackend has been introduced to build
IPointReaders, IPointWriters and IPointRecordWriters from BinaryIOs and
LasHeaders.

The code relative to reading, appending and writing has been moved to a
unique module for each backend (lazrs and laszip).

For each backend, the corresponding ILazBackend has been implemented.
LazBackend.detect_available() now uses _DEFAULT_BACKENDS

The code of lasreader, lasappender and laswriter has been adapted to
make use of the ILazBackend interface, and is no longer directly coupled
to laszip or lazrs.

Some modules have been split in order to avoid circular imports
(laspy.compression for example)

- Renamed Laszip and Lazrs to LaszipBackend and LazrsBackend
- Remove unused abstractmethods of ILazBackend
- Moved IPointWriter to _pointwriter.py
- Moved IPointRecordWriter to _pointrecordwriter.py
- Moved LazrsPointWriter to _lazrsbackend.py
- Moved LaszipPointWriter to _laszipbackend.py
- Added create_writer methods to all ILazBackends
- ILazBackend.create_writer is now used in laswriter.py
- Renamed IPointRecordWriter to IPointAppender
