+++ laspy/laswriter.py
@@ -1,31 +1,16 @@
-import abc
-import io
 import logging
 from copy import deepcopy
 from typing import BinaryIO, Optional, Union, Iterable
 
-import numpy as np
-
+from ._pointwriter import IPointWriter
 from .compression import LazBackend
 from .errors import LaspyException
 from .header import LasHeader
 from .point import dims
-from .point.format import PointFormat
 from .point.record import PackedPointRecord
-from .vlrs.known import LasZipVlr
 from .vlrs.vlrlist import VLRList
 
 logger = logging.getLogger(__name__)
-
-try:
-    import lazrs
-except ModuleNotFoundError:
-    pass
-
-try:
-    import laszip
-except ModuleNotFoundError:
-    pass
 
 
 class LasWriter:
@@ -198,18 +183,7 @@ class LasWriter:
                 if not backend.is_available():
                     raise LaspyException(f"The '{backend}' is not available")
 
-                if backend == LazBackend.Laszip:
-                    return LaszipPointWriter(self.dest, self.header)
-                elif backend == LazBackend.LazrsParallel:
-                    return LazrsPointWriter(
-                        self.dest, self.header.point_format, parallel=True
-                    )
-                elif backend == LazBackend.Lazrs:
-                    return LazrsPointWriter(
-                        self.dest, self.header.point_format, parallel=False
-                    )
-                else:
-                    raise LaspyException("Unknown LazBacked: {}".format(backend))
+                return backend.create_writer(self.dest, self.header)
             except Exception as e:
                 logger.error(e)
                 last_error = e
@@ -224,37 +198,6 @@ class LasWriter:
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.close()
-
-
-class IPointWriter(abc.ABC):
-    """Interface to be implemented by the actual
-    PointWriter backend
-
-    """
-
-    @property
-    @abc.abstractmethod
-    def destination(self) -> BinaryIO:
-        ...
-
-    def write_initial_header_and_vlrs(
-        self, header: LasHeader, encoding_errors: str
-    ) -> None:
-        header.write_to(self.destination, encoding_errors=encoding_errors)
-
-    @abc.abstractmethod
-    def write_points(self, points: PackedPointRecord) -> None:
-        ...
-
-    @abc.abstractmethod
-    def done(self) -> None:
-        ...
-
-    def write_updated_header(self, header: LasHeader, encoding_errors: str):
-        self.destination.seek(0, io.SEEK_SET)
-        header.write_to(
-            self.destination, ensure_same_size=True, encoding_errors=encoding_errors
-        )
 
 
 class UncompressedPointWriter(IPointWriter):
@@ -274,101 +217,3 @@ class UncompressedPointWriter(IPointWriter):
 
     def done(self) -> None:
         pass
-
-
-class LaszipPointWriter(IPointWriter):
-    """
-    Compressed point writer using laszip backend
-    """
-
-    def __init__(self, dest: BinaryIO, header: LasHeader) -> None:
-        self.dest = dest
-        header.set_compressed(False)
-        with io.BytesIO() as tmp:
-            header.write_to(tmp)
-            header_bytes = tmp.getvalue()
-
-        self.zipper = laszip.LasZipper(self.dest, header_bytes)
-        zipper_header = self.zipper.header
-        assert zipper_header.point_data_format == header.point_format.id
-        assert zipper_header.point_data_record_length == header.point_format.size
-
-        header.set_compressed(True)
-
-    @property
-    def destination(self) -> BinaryIO:
-        return self.dest
-
-    def write_points(self, points: PackedPointRecord) -> None:
-        points_bytes = np.frombuffer(points.array, np.uint8)
-        self.zipper.compress(points_bytes)
-
-    def done(self) -> None:
-        self.zipper.done()
-
-    def write_initial_header_and_vlrs(
-        self, header: LasHeader, encoding_errors: str
-    ) -> None:
-        # Do nothing as creating the laszip zipper writes the header and vlrs
-        pass
-
-    def write_updated_header(self, header: LasHeader, encoding_errors: str) -> None:
-        if header.number_of_evlrs != 0:
-            # We wrote some evlrs, we have to update the header
-            self.dest.seek(0, io.SEEK_SET)
-            file_header = LasHeader.read_from(self.dest)
-            end_of_header_pos = self.dest.tell()
-            file_header.number_of_evlrs = header.number_of_evlrs
-            file_header.start_of_first_evlr = header.start_of_first_evlr
-            self.dest.seek(0, io.SEEK_SET)
-            file_header.write_to(
-                self.dest, ensure_same_size=True, encoding_errors=encoding_errors
-            )
-            assert self.dest.tell() == end_of_header_pos
-
-
-class LazrsPointWriter(IPointWriter):
-    """
-    Compressed point writer using lazrs backend
-    """
-
-    def __init__(
-        self, dest: BinaryIO, point_format: PointFormat, parallel: bool
-    ) -> None:
-        self.dest = dest
-        self.vlr = lazrs.LazVlr.new_for_compression(
-            point_format.id, point_format.num_extra_bytes
-        )
-        self.parallel = parallel
-        self.compressor: Optional[
-            Union[lazrs.ParLasZipCompressor, lazrs.LasZipCompressor]
-        ] = None
-
-    def write_initial_header_and_vlrs(
-        self, header: LasHeader, encoding_errors: str
-    ) -> None:
-        laszip_vlr = LasZipVlr(self.vlr.record_data())
-        header.vlrs.append(laszip_vlr)
-        super().write_initial_header_and_vlrs(header, encoding_errors)
-        # We have to initialize our compressor here
-        # because on init, it writes the offset to chunk table
-        # so the header and vlrs have to be written
-        if self.parallel:
-            self.compressor = lazrs.ParLasZipCompressor(self.dest, self.vlr)
-        else:
-            self.compressor = lazrs.LasZipCompressor(self.dest, self.vlr)
-
-    @property
-    def destination(self) -> BinaryIO:
-        return self.dest
-
-    def write_points(self, points: PackedPointRecord) -> None:
-        assert (
-            self.compressor is not None
-        ), "Trying to write points without having written header"
-        points_bytes = np.frombuffer(points.array, np.uint8)
-        self.compressor.compress_many(points_bytes)
-
-    def done(self) -> None:
-        if self.compressor is not None:
-            self.compressor.done()
